Add support for subtarget
Subtargets allow the customization for different use-cases of the processing chain.
To have a slightly different behaviour for a cosmo-run for the project 'FOO',
extend the Subclass-enum and the string_to_enum-dict in jobs/tools/__init__.py.
Then, when you add the variable `subtarget = 'FOO'` to your config.py, the
processing chain can do different stuff for the 'FOO'-subtarget by checking
    `>>> (cfg.target is tools.Target.COSMO and`
    `...  cfg.target.subtarget is tools.Subtarget.CHE)`
    `True`